Practica 3, 5 y 6 lista. Con bottom navigation y listo el diseño del perfil Errores arreglados

+++ app/src/main/java/com/alejandrazuleta/clase2/model/cursoinscrito.kt
@@ -8,5 +8,5 @@ import java.io.Serializable
 class cursoinscrito(
     @ColumnInfo(name = "idusuario") val idusuario: String = "",
     @ColumnInfo(name = "idcurso") val idcurso: String = "",
-    @ColumnInfo(name = "notas") val notas: List<Int> = arrayListOf<Int>()
+    @ColumnInfo(name = "notas") val notas: List<Double> = arrayListOf<Double>()
 ): Serializable
